+++ packages/cli/src/cmds/index/IndexResult.ts
@@ -1,3 +1,7 @@
+import type { Metadata } from '@appland/models';
+
 export type IndexResult = {
   error?: Error;
+  metadata: Metadata;
+  numEvents: number;
 };