+++ packages/cli/src/fingerprint/fingerprintDirectoryCommand.ts
@@ -35,8 +35,8 @@ class FingerprintDirectoryCommand {
   }
 
   async files(fn) {
-    return findFiles(this.directory, '*.appmap.json', fn);
+    return findFiles(this.directory, '.appmap.json', fn);
   }
 }
 
-module.exports = FingerprintDirectoryCommand;
+export default FingerprintDirectoryCommand;