+++ packages/cli/tests/unit/fingerprint/fingerprinter.spec.ts
@@ -1,8 +1,10 @@
-import { mkdtemp, open, rm } from 'fs/promises';
+import { mkdtemp, open, rm, writeFile } from 'fs/promises';
 import path from 'node:path';
 import { tmpdir } from 'os';
 import FileTooLargeError from '../../../src/fingerprint/fileTooLargeError';
 import Fingerprinter from '../../../src/fingerprint/fingerprinter';
+import * as utils from '../../../src/utils';
+import { relative, resolve } from 'path';
 
 function withTempDir<T>(closure: (dir: string) => Promise<T>): () => Promise<T> {
   return async function () {
@@ -25,6 +27,24 @@ describe(Fingerprinter, () => {
       await file.close();
       const fingerprinter = new Fingerprinter();
       await expect(fingerprinter.fingerprint(filePath)).rejects.toThrow(FileTooLargeError);
+    })
+  );
+
+  it(
+    'prints indexed events when verbose',
+    withTempDir(async (dir) => {
+      expect.assertions(1);
+
+      const filePath = path.join(dir, 'test.appmap.json');
+      await writeFile(filePath, JSON.stringify({ metadata: {} }));
+
+      jest.spyOn(utils, 'verbose').mockReturnValue(true);
+      const consoleSpy = jest.spyOn(console, 'log');
+
+      // use relative path to test that the message resolves absolute back
+      await new Fingerprinter().fingerprint(relative(process.cwd(), filePath));
+
+      expect(consoleSpy.mock.calls).toContainEqual([`Indexed ${resolve(filePath)}`]);
     })
   );
 });