+++ packages/components/src/components/sequence/CallAction.vue
@@ -441,6 +441,10 @@ export default {
   fill: $sequence-call-line-color;
 }
 
+.connecting-span {
+  position: relative;
+}
+
 .call-right {
   .single-span {
     left: calc(((var(--caller-lifecycle-depth)) * $sequence-activation-gutter-width));
@@ -470,6 +474,7 @@ export default {
   .connecting-span {
     border-radius: 0;
     border-top: 4px solid $black;
+    position: relative;
   }
 
   .arrow {
@@ -487,6 +492,7 @@ export default {
   .connecting-span {
     border-radius: 0;
     border-top: 4px solid $black;
+    position: relative;
   }
 
   .arrow-head {
@@ -512,8 +518,11 @@ export default {
   background-color: black;
   border: 1px solid gray;
   position: relative;
-  height: calc(100% - 6px - (var(--open-group-count) + var(--close-group-count)) * 40px);
-  top: calc(var(--open-group-count) * 40px + 35px);
+  height: calc(
+    100% - 6px - (var(--open-group-count) + var(--close-group-count)) * 40px -
+      var(--first-lifeline-first-action) * 80px
+  );
+  top: calc(var(--open-group-count) * 40px + 35px + var(--first-lifeline-first-action) * 80px);
   left: calc(100% - (var(--callee-lifecycle-depth) * $sequence-activation-gutter-width));
 }
 </style>