+++ packages/components/src/stories/DiagramSequence.stories.js
@@ -1,5 +1,5 @@
 import VDiagramSequence from '@/components/DiagramSequence.vue';
-import appland_api_key_diff from '@/stories/data/sequence/appland_api_key_diff.sequence.json';
+import micropost_diff from '@/stories/data/sequence/Users_profile_profile_display.diff.sequence.json';
 import create_api_key from '@/stories/data/sequence/create_api_key.sequence.json';
 import list_users from '@/stories/data/sequence/list_users.sequence.json';
 import list_users_prefetch from '@/stories/data/sequence/list_users_prefetch.sequence.json';
@@ -17,9 +17,9 @@ const Template = (args, { argTypes }) => ({
   template: '<v-diagram-sequence v-bind="$props"/>',
 });
 
-export const AppLandApiKeyDiff = Template.bind({});
-AppLandApiKeyDiff.args = {
-  serializedDiagram: appland_api_key_diff,
+export const MicropostUserProfileDiff = Template.bind({});
+MicropostUserProfileDiff.args = {
+  serializedDiagram: micropost_diff,
 };
 export const Empty = Template.bind({});