+++ packages/components/src/stories/VsCodeExtension.stories.js
@@ -7,6 +7,10 @@ import petClinicScenario from './data/java_scenario.json';
 import diffScenario from './data/diff_base.json';
 import appland1 from './data/Application_page_component_diagram_highlights_node_connections_upon_selection.appmap.json';
 import appland2 from './data/ApplicationsController_scenarios_list_when_the_user_is_anonymous_is_not_found.appmap.json';
+import longCallLabels from './data/long_call_labels_are_cut_off.appmap.json';
+import diffAppMap from './data/sequence-diff/Users_profile_profile_display.appmap.json';
+import diffSequenceDiagram from './data/sequence-diff/Users_profile_profile_display.diff.sequence.json';
+import longPackage from './data/long-package.appmap.json';
 import mapWithFindings from './data/appmap_with_finding.json';
 import mapWithTwoFindings from './data/appmap_with_two_findings.json';
 import patchNotes from './data/patch_notes_html';
@@ -21,8 +25,15 @@ const scenarioData = {
   'pet-clinic': petClinicScenario,
   appland1,
   appland2,
+  longCallLabels,
+  mapWithDiff: diffAppMap,
+  longPackage,
   mapWithFindings,
   mapWithTwoFindings,
+};
+
+const sequenceDiagramData = {
+  mapWithDiff: diffSequenceDiagram,
 };
 
 export default {
@@ -55,8 +66,9 @@ const Template = (args, { argTypes }) => ({
   template: '<v-vs-code-extension v-bind="$props" ref="vsCode" />',
   mounted() {
     const scenario = scenarioData[args.scenario];
+    const sequenceDiagram = sequenceDiagramData[args.scenario];
     if (scenario) {
-      this.$refs.vsCode.loadData(scenario);
+      this.$refs.vsCode.loadData(scenario, sequenceDiagram);
     }
 
     bindResolvePath(this);
@@ -68,6 +80,13 @@ export const extension = Template.bind({});
 export const extensionWithDefaultSequenceView = Template.bind({});
 extensionWithDefaultSequenceView.args = {
   defaultView: VIEW_SEQUENCE,
+};
+
+export const extensionWithSequenceDiff = Template.bind({});
+extensionWithSequenceDiff.args = {
+  defaultView: VIEW_SEQUENCE,
+  scenario: 'mapWithDiff',
+  interactive: false,
 };
 
 export const extensionWithSavedFilters = Template.bind({});